+++ src/components/UploadImageWidget/uploadWidgetReducer.ts
