--- src/lib/utils/index.ts
@@ -1,1 +1,0 @@
-export { default as uploadImageToCloudinary } from "./uploadImageToCloudinary";
