--- src/views/test/reducers/index.ts
@@ -1,4 +1,0 @@
-export { default as uploadWidgetReducer } from "./uploadWidgetReducer";
-export {
-  default as uploadToCloudinaryReducer,
-} from "./uploadToCloudinaryReducer";
